Added RandomFactionCamouflage
